Account functionality and log-in menu
Added log-in menu and created CPP base class for menu. It handles checking the users entered data against the database to see if the user exists. If it does and the password is correct, start the game. If the user exists but the password is incorrect, notify user of incorrect password. If user does not exist, create user in database with provided password.